+++ lib/auth-backend/venue_identity.rb
@@ -4,7 +4,7 @@ module Auth::Backend
   class VenueIdentity < ::ActiveRecord::Base
     attr_accessible :venue, :venue_id, :user_id, :name, :email
 
-    validates :venue, inclusion: {in: %w{facebook galaxy-spiral}}
+    validates :venue, inclusion: {in: %w{facebook spiral-galaxy}}
 
     belongs_to :user
   end